requirements.txt
requests>=2.25.1
Flask>=1.1.2
Flask-Cors>=3.0.10
python-dotenv>=0.15.0
Flask-Dance>=3.0.0
